Split offset.test.js into offsetGeom2.test.js and offsetPath2.test.js
